data: normalize employee department and designation relations

+++ EduOS.Core/Entities/Employees/Employee.cs
@@ -20,8 +20,8 @@ namespace EduOS.Core.Entities.Employees
         public string? Email { get; set; }
         public string? Address { get; set; }
         public string? NID { get; set; }
-        public int DesignationId { get; set; }
-        public int? DepartmentId { get; set; }
+        public long DesignationId { get; set; }
+        public long? DepartmentId { get; set; }
         public DateTime JoiningDate { get; set; }
         public decimal Salary { get; set; }
         public string? PhotoUrl { get; set; }
